finished adding unit tests and tweaked the getMilesValue method

+++ src/main/java/RewardValue.java
@@ -15,7 +15,7 @@ public class RewardValue {
         return cashValue;
     }
 
-    public double getMilesValue() {
-        return this.cashValue / milesToCashRate;
+    public int getMilesValue() {
+        return (int) (this.cashValue / milesToCashRate);
     }
 }
